upstream: Unify integer type for EClient::m_port and arguments as discussed in PR #667

+++ cpp/client/EClientSocket.cpp
@@ -53,7 +53,7 @@ void EClientSocket::asyncEConnect(bool val) {
     m_asyncEConnect = val;
 }
 
-bool EClientSocket::eConnect( const char *host, unsigned int port, int clientId, bool extraAuth)
+bool EClientSocket::eConnect( const char *host, int port, int clientId, bool extraAuth)
 {
 	if( m_fd == -2) {
 		getWrapper()->error( NO_VALID_ID, FAIL_CREATE_SOCK.code(), FAIL_CREATE_SOCK.msg());
@@ -279,7 +279,7 @@ void EClientSocket::serverVersion(int version, const char *time) {
 void EClientSocket::redirect(const char *host, int port) {
 	const char* hostNorm = (host && *host) ? host : "127.0.0.1";
 
-	if( (hostNorm != this->host() || (port > 0 && (unsigned int)port != this->port()))) {
+	if( (hostNorm != this->host() || (port > 0 && port != this->port()))) {
         if (!m_allowRedirect) {
             getWrapper()->error(NO_VALID_ID, CONNECT_FAIL.code(), CONNECT_FAIL.msg());
